requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
Pillow>=10.0.0
opencv-python-headless>=4.8.0
numpy>=1.24.0
scikit-learn>=1.3.0
scipy>=1.11.0
vtracer>=0.6.0
cairosvg>=2.7.0
svgwrite>=1.4.3
svgpathtools>=1.6.0
ezdxf>=0.18.0
aiofiles>=23.0.0
jinja2>=3.1.0
